Reflection: RdRpc interface in RdExt attribute
Add a possiblity to explicitly set RdRpc interface not only from
interface declaration, but from RdExt itself. It can be used for
inversion of control pattern. RdRpc interface should follow several
rules, but it can just be a regular interface, which can be placed in
core part of the product without dependency to RdFramework.