Assertion review
The most of the assertions were disabled in release build previously
using JET_MODE_ASSERT define. The code was written with the assumption
that Assertion.Assert methods invocations are removed in release mode.

This is no longer truth. The NuGet packages builds with the
JET_MODE_ASSERT enabled. To reduce performance impact of the Assertion
I added additional checks for Mode.Assertion to help JIT to remove
the branch with the check (in perf-critical code)

Additionally, in non-performance critical places, I converted
Assertions.Assert to Assertion.Require method, which is not marked by
ConditionalCompilation.