Built-in serializers rework
Convert the process of finding built-in serializer into two-stage
process: detection and creation. Defining all possible built-in
serializer types makes getting required member straightforward and
allows to provide a meaningfull error message if something went wrong.

It is now possible to show which exactly signature is missing. Also, the
number of calls to GetMembers(BindingFlags) has been reduced, so it
should reduce memory traffic when allocating an array to return.