rd-kt: fix partial read behavior in SocketWire
Before this change, the following was theoretically possible: we read a package body, then disconnect happens in the middle, and we end up with partial body filled with zeros. We'll try to interpret it as valid payload instead of handling the disconnect. Disconnect will only be detected when we try reading the next package.